Issue #81: Allow columns to be hidden on search results
Click on header of advanced search results table hides column. Index
of all hidden columns are kept in the array and link for displaying
all columns is displayed depending of array content. If array is empty
link is hidden; otherwise it is shown. When link is clicked array is
emptied and all columns are shown.