Fixed duplicate calls and ids
The read cookie was applying the bulk selected to all rows with the Id meaning that the result rows that don't have a checkbox in a parent child relationship
had bulk_selected = true.  The new read method only applies the bulk_selected flag to the first instance of each ID.  Also, removed watch statement that wasn't
needed anymore.