Send "No attachment" note if an attachment is blank. #T1072
I had to add some mocks throughout the testing file because apparently
the function that I made doesn't play very nicely with the Tempfile class.  This
seemed like it was in keeping with the pattern of other tests, but let me know
if you have a suggestion of how to improve that.

* Update OpenMailer with blank_attachment? function
* Update attachment creation method to handle blank and oversized attachments
* Add tests for blank attachments