Fixing Jeremy's bug.  See extended commit message.
One would think that in the 10+ years that Jeremy has been doing web development
that he would have learned how a <select multiple> tag is interpreted by the
browser, but apparently not.  I'm now searching the codebase to make sure that
all of our <blink> tags are properly closed and that all of our image maps have
fully enclosed polygons.