Sow 1414 new business rules (#2260)
* More annotation madness.

* Adding in the ability to do gt, lt, eq, and nq on two fields.

I finally settled on turning this into an equality check vs a relative equality check. This class is already getting messy and I didn't want to make it even more messy than it is. Combine that with having to deal with the default error message and this just makes more sense.

* Checking for blank? as opposed to respond_to?