Shipments tracking datetime fix (#3321)
* Remove duplicate input fields

It's not clear to me why the same input fields were added in the rails upgrade here. The ng-model points to the same variables.

* Fix cancellation message

- Fix spelling mistakes
- Make appearance more consistent with the rest of the application
- Move undo button to inside the message box

* Format datetimes nicely

* Fix datetime input fields

Datetime-local input types handle the datetime more elegantly than trying to parse the time from date like for shp_in_warehouse_time. It however does need to be a Date object. Parsing each date field as a date on load is probably not the best solution here.

* Ensure that datetime values are set or unset

- Make sure that any datetimes are set appropriately or are unset as we don't want fields to be populated with the current datetime by accident when it actually isn't populated
- Also make the spelling of 'cancelled' consistent across the page